requirements.txt
# FastAPI 框架和 Uvicorn 伺服器
fastapi==0.129.0
uvicorn==0.41.0
starlette==0.52.1

# LINE Bot SDK
line-bot-sdk==3.22.0

# 環境變數設定
python-dotenv==1.1.1

# HTTP 請求相關（用於 Gemini API 和其他 HTTP 請求）
requests==2.32.5
httpx==0.28.1

# 測試相關（pytest-asyncio 0.25 僅支援 pytest<9）
pytest>=8.2,<9
pytest-asyncio==0.25.0
pytest-cov==7.0.0

# 工具套件
pydantic==2.12.5
typing-extensions==4.15.0